Tue, Jul 9, 2024, 5:47 PM +02:00

+++ 3.Exceptions/fuel/fuel.py
@@ -15,11 +15,11 @@ def userfraction():
             fraction = input("Fraction: ")
             # Split l'input en X/Y
             x, y = fraction.split("/")
-            x = float(x)
-            y = float(y)
+            x = int(x)
+            y = int(y)
 
             # Convertir la fraction en pourcentage arrondi à l'intégral le plus proche
-            percentage = int(round((x*100)/y))
+            percentage = int(round((x*100)/y)) 
 
 
             # Boucler et redemander un input si l'input est supérieur à 100% ou en cas de  ZeroDivisionError et ValueError
